+++ core/terraform/outputs.sh
@@ -30,12 +30,9 @@ if [ -f ../.env ]; then
   source ../.env
 fi
 
-# Add a few extra values to the file to help us (i.e. for local debugging api_app and resource processor)
-# shellcheck disable=SC2129
-echo "TEST_WORKSPACE_APP_ID='${WORKSPACE_API_CLIENT_ID}'" >> ../private.env
-echo "TEST_WORKSPACE_APP_SECRET='${WORKSPACE_API_CLIENT_SECRET}'" >> ../private.env
-
 # These next ones from Check Dependencies
-echo "SUBSCRIPTION_ID='${SUB_ID}'" >> ../private.env
-echo "AZURE_SUBSCRIPTION_ID='${SUB_ID}'" >> ../private.env
-echo "AZURE_TENANT_ID='${TENANT_ID}'" >> ../private.env
+{
+  echo "SUBSCRIPTION_ID='${SUB_ID}'"
+  echo "AZURE_SUBSCRIPTION_ID='${SUB_ID}'"
+  echo "AZURE_TENANT_ID='${TENANT_ID}'"
+} >> ../private.env