+++ e2e_tests/config.py
@@ -16,8 +16,6 @@ TEST_APP_ID: str = config("TEST_APP_ID", default="")
 AAD_TENANT_ID: str = config("AAD_TENANT_ID", default="")
 TEST_ACCOUNT_CLIENT_ID: str = config("TEST_ACCOUNT_CLIENT_ID", default="")
 TEST_ACCOUNT_CLIENT_SECRET: str = config("TEST_ACCOUNT_CLIENT_SECRET", default="")
-TEST_WORKSPACE_APP_ID: str = config("TEST_WORKSPACE_APP_ID", default="")
-TEST_WORKSPACE_APP_SECRET: str = config("TEST_WORKSPACE_APP_SECRET", default="")
 TEST_WORKSPACE_APP_PLAN: str = config("WORKSPACE_APP_SERVICE_PLAN_SKU", default="")
 
 # Set workspace id of an existing workspace to skip creation of a workspace during E2E tests