+++ templates/workspaces/base/terraform/aad/aad.tf
@@ -112,8 +112,9 @@ resource "azuread_service_principal_delegated_permission_grant" "ui" {
   claim_values                         = ["user_impersonation"]
 }
 
-resource "azuread_service_principal_password" "workspace" {
-  service_principal_id = azuread_service_principal.workspace.id
+resource "azuread_application_password" "workspace" {
+  application_id = azuread_application.workspace.id
+  display_name   = "Workspace Password"
 }
 
 resource "azurerm_key_vault_secret" "client_id" {
@@ -127,7 +128,7 @@ resource "azurerm_key_vault_secret" "client_id" {
 
 resource "azurerm_key_vault_secret" "client_secret" {
   name         = "workspace-client-secret"
-  value        = azuread_service_principal_password.workspace.value
+  value        = azuread_application_password.workspace.value
   key_vault_id = var.key_vault_id
   tags         = var.tre_workspace_tags
 
@@ -135,7 +136,7 @@ resource "azurerm_key_vault_secret" "client_secret" {
 }
 
 resource "azuread_app_role_assignment" "workspace_owner" {
-  app_role_id         = azuread_service_principal.workspace.app_role_ids["WorkspaceOwner"]
+  app_role_id         = azuread_application.workspace.app_role_ids["WorkspaceOwner"]
   principal_object_id = var.workspace_owner_object_id
   resource_object_id  = azuread_service_principal.workspace.object_id
 }
@@ -169,21 +170,21 @@ resource "azuread_group_member" "workspace_owner" {
 
 resource "azuread_app_role_assignment" "workspace_owners_group" {
   count               = var.create_aad_groups ? 1 : 0
-  app_role_id         = azuread_service_principal.workspace.app_role_ids["WorkspaceOwner"]
+  app_role_id         = azuread_application.workspace.app_role_ids["WorkspaceOwner"]
   principal_object_id = azuread_group.workspace_owners[count.index].object_id
   resource_object_id  = azuread_service_principal.workspace.object_id
 }
 
 resource "azuread_app_role_assignment" "workspace_researchers_group" {
   count               = var.create_aad_groups ? 1 : 0
-  app_role_id         = azuread_service_principal.workspace.app_role_ids["WorkspaceResearcher"]
+  app_role_id         = azuread_application.workspace.app_role_ids["WorkspaceResearcher"]
   principal_object_id = azuread_group.workspace_researchers[count.index].object_id
   resource_object_id  = azuread_service_principal.workspace.object_id
 }
 
 resource "azuread_app_role_assignment" "workspace_airlock_managers_group" {
   count               = var.create_aad_groups ? 1 : 0
-  app_role_id         = azuread_service_principal.workspace.app_role_ids["AirlockManager"]
+  app_role_id         = azuread_application.workspace.app_role_ids["AirlockManager"]
   principal_object_id = azuread_group.workspace_airlock_managers[count.index].object_id
   resource_object_id  = azuread_service_principal.workspace.object_id
 }