+++ templates/workspaces/base/terraform/azure-monitor/azure-monitor.tf
@@ -194,6 +194,7 @@ resource "azurerm_private_endpoint" "azure_monitor_private_endpoint" {
 
   depends_on = [
     azurerm_monitor_private_link_scoped_service.ampls_app_insights,
+    azurerm_monitor_private_link_scoped_service.ampls_log_anaytics,
   ]
 }